pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-crm"
version = "0.1.0"
description = "A small CRM application built with FastAPI and SQLite"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "alembic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "jinja2",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
